streamboard_backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django_cors_headers==4.7.0
drf_yasg==1.21.10
pillow==11.2.1